fix: param server test, refactor
